Migrate RaceDB to LuaLazyCache
