--- runtime/data/race.lua
@@ -1,2312 +1,0 @@
-race.kobold = {
-  is_extra = true,
-  ordering = 20010,
-  breed_power = 250,
-  male_image = 171,
-  female_image = 171,
-  min_age = 15,
-  max_age = 44,
-  height = 150,
-  male_ratio = 50,
-  skills = {
-    _2 = 90,
-    _3 = 80,
-    _10 = 8,
-    _11 = 6,
-    _12 = 5,
-    _13 = 3,
-    _14 = 2,
-    _15 = 2,
-    _16 = 4,
-    _17 = 2,
-    _18 = 70,
-    _106 = 2,
-    _300 = 3,
-    _157 = 2,
-    _174 = 3,
-  },
-  body_parts = {1, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.orc = {
-  is_extra = true,
-  ordering = 20020,
-  breed_power = 300,
-  male_image = 165,
-  female_image = 165,
-  min_age = 15,
-  max_age = 44,
-  height = 150,
-  male_ratio = 51,
-  skills = {
-    _2 = 130,
-    _3 = 80,
-    _10 = 5,
-    _11 = 6,
-    _12 = 3,
-    _13 = 2,
-    _14 = 2,
-    _15 = 3,
-    _16 = 4,
-    _17 = 2,
-    _18 = 70,
-    _106 = 1,
-    _154 = 3,
-    _169 = 3,
-    _168 = 3,
-  },
-  body_parts = {1, 4, 3, 5, 5, 6, 7, 8, 9},
-}
-race.troll = {
-  is_extra = true,
-  ordering = 20030,
-  breed_power = 250,
-  male_image = 391,
-  female_image = 391,
-  min_age = 15,
-  max_age = 44,
-  height = 400,
-  male_ratio = 51,
-  pv_correction = 150,
-  skills = {
-    _2 = 90,
-    _3 = 80,
-    _10 = 10,
-    _11 = 9,
-    _12 = 1,
-    _13 = 2,
-    _14 = 2,
-    _15 = 2,
-    _16 = 3,
-    _17 = 3,
-    _18 = 60,
-    _106 = 3,
-    _154 = 40,
-  },
-  body_parts = {1, 4, 3, 5, 5, 7, 8},
-}
-race.lizardman = {
-  is_extra = true,
-  ordering = 20040,
-  breed_power = 300,
-  male_image = 397,
-  female_image = 397,
-  min_age = 15,
-  max_age = 44,
-  height = 240,
-  male_ratio = 51,
-  dv_correction = 120,
-  skills = {
-    _2 = 120,
-    _3 = 70,
-    _10 = 7,
-    _11 = 5,
-    _12 = 6,
-    _13 = 7,
-    _14 = 4,
-    _15 = 4,
-    _16 = 5,
-    _17 = 3,
-    _18 = 80,
-    _106 = 2,
-    _104 = 3,
-    _168 = 3,
-    _173 = 2,
-  },
-  body_parts = {1, 4, 3, 5, 5, 6, 7, 8, 9},
-}
-race.minotaur = {
-  is_extra = true,
-  ordering = 20050,
-  breed_power = 300,
-  male_image = 398,
-  female_image = 398,
-  min_age = 15,
-  max_age = 44,
-  height = 350,
-  male_ratio = 51,
-  skills = {
-    _2 = 160,
-    _3 = 60,
-    _10 = 12,
-    _11 = 8,
-    _12 = 3,
-    _13 = 4,
-    _14 = 3,
-    _15 = 4,
-    _16 = 7,
-    _17 = 4,
-    _18 = 65,
-    _106 = 3,
-    _152 = 4,
-    _186 = 3,
-  },
-  body_parts = {1, 4, 5, 5, 6, 8, 9},
-}
-race.yerles = {
-  is_extra = false,
-  ordering = 10010,
-  breed_power = 220,
-  male_image = 1,
-  female_image = 2,
-  min_age = 15,
-  max_age = 34,
-  height = 165,
-  male_ratio = 52,
-  skills = {
-    _2 = 110,
-    _3 = 90,
-    _10 = 7,
-    _11 = 8,
-    _12 = 8,
-    _13 = 8,
-    _14 = 12,
-    _15 = 8,
-    _16 = 8,
-    _17 = 7,
-    _18 = 70,
-    _106 = 2,
-    _110 = 5,
-    _150 = 3,
-    _156 = 2,
-    _111 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.norland = {
-  is_extra = true,
-  ordering = 20060,
-  breed_power = 220,
-  male_image = 1,
-  female_image = 2,
-  min_age = 15,
-  max_age = 34,
-  height = 170,
-  male_ratio = 52,
-  skills = {
-    _2 = 100,
-    _3 = 90,
-    _10 = 8,
-    _11 = 8,
-    _12 = 6,
-    _13 = 6,
-    _14 = 7,
-    _15 = 9,
-    _16 = 7,
-    _17 = 6,
-    _18 = 70,
-    _106 = 2,
-    _172 = 3,
-    _152 = 3,
-    _167 = 3,
-    _188 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.eulderna = {
-  is_extra = false,
-  ordering = 10020,
-  breed_power = 180,
-  male_image = 5,
-  female_image = 6,
-  min_age = 16,
-  max_age = 35,
-  height = 175,
-  male_ratio = 52,
-  skills = {
-    _2 = 100,
-    _3 = 100,
-    _10 = 6,
-    _11 = 7,
-    _12 = 7,
-    _13 = 10,
-    _14 = 8,
-    _15 = 10,
-    _16 = 12,
-    _17 = 8,
-    _18 = 70,
-    _106 = 2,
-    _172 = 5,
-    _150 = 3,
-    _174 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.fairy = {
-  is_extra = false,
-  ordering = 10030,
-  breed_power = 180,
-  male_image = 390,
-  female_image = 390,
-  min_age = 5,
-  max_age = 104,
-  height = 50,
-  male_ratio = 52,
-  resistances = {
-    _60 = 200,
-    _53 = 200,
-    _58 = 200,
-    _56 = 200,
-    _54 = 200,
-    _57 = 200,
-    _59 = 200,
-  },
-  dv_correction = 250,
-  skills = {
-    _2 = 40,
-    _3 = 130,
-    _10 = 3,
-    _11 = 4,
-    _12 = 9,
-    _13 = 10,
-    _14 = 8,
-    _15 = 8,
-    _16 = 13,
-    _17 = 12,
-    _18 = 120,
-    _106 = 1,
-    _172 = 5,
-    _300 = 3,
-    _171 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.asura = {
-  is_extra = true,
-  ordering = 20070,
-  breed_power = 100,
-  male_image = 405,
-  female_image = 405,
-  min_age = 15,
-  max_age = 34,
-  height = 220,
-  male_ratio = 52,
-  dv_correction = 200,
-  skills = {
-    _2 = 90,
-    _3 = 70,
-    _10 = 11,
-    _11 = 9,
-    _12 = 14,
-    _13 = 6,
-    _14 = 4,
-    _15 = 4,
-    _16 = 5,
-    _17 = 5,
-    _18 = 70,
-    _106 = 2,
-    _166 = 30,
-    _187 = 6,
-    _161 = 4,
-  },
-  body_parts = {5, 5, 5, 5, 2},
-}
-race.slime = {
-  is_extra = true,
-  ordering = 20080,
-  breed_power = 700,
-  male_image = 168,
-  female_image = 168,
-  special_attack_type = 2,
-  min_age = 1,
-  max_age = 10,
-  height = 40,
-  male_ratio = 54,
-  skills = {
-    _2 = 80,
-    _3 = 100,
-    _10 = 4,
-    _11 = 5,
-    _12 = 7,
-    _13 = 5,
-    _14 = 6,
-    _15 = 8,
-    _16 = 4,
-    _17 = 13,
-    _18 = 55,
-    _106 = 2,
-    _173 = 2,
-    _183 = 3,
-  },
-  body_parts = {1},
-}
-race.wolf = {
-  is_extra = true,
-  ordering = 20090,
-  breed_power = 800,
-  male_image = 254,
-  female_image = 254,
-  min_age = 2,
-  max_age = 11,
-  height = 100,
-  male_ratio = 55,
-  dv_correction = 140,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 8,
-    _11 = 8,
-    _12 = 6,
-    _13 = 11,
-    _14 = 7,
-    _15 = 8,
-    _16 = 4,
-    _17 = 6,
-    _18 = 80,
-    _106 = 2,
-    _173 = 2,
-    _187 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 7, 9},
-}
-race.dwarf = {
-  is_extra = false,
-  ordering = 10040,
-  breed_power = 150,
-  male_image = 66,
-  female_image = 66,
-  min_age = 20,
-  max_age = 79,
-  height = 100,
-  male_ratio = 56,
-  skills = {
-    _2 = 100,
-    _3 = 100,
-    _10 = 10,
-    _11 = 11,
-    _12 = 8,
-    _13 = 6,
-    _14 = 6,
-    _15 = 7,
-    _16 = 6,
-    _17 = 6,
-    _18 = 70,
-    _106 = 2,
-    _184 = 4,
-    _179 = 3,
-    _163 = 4,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.juere = {
-  is_extra = false,
-  ordering = 10050,
-  breed_power = 210,
-  male_image = 9,
-  female_image = 10,
-  min_age = 15,
-  max_age = 44,
-  height = 165,
-  male_ratio = 50,
-  skills = {
-    _2 = 100,
-    _3 = 100,
-    _10 = 7,
-    _11 = 8,
-    _12 = 12,
-    _13 = 10,
-    _14 = 10,
-    _15 = 8,
-    _16 = 8,
-    _17 = 11,
-    _18 = 70,
-    _106 = 2,
-    _183 = 4,
-    _158 = 2,
-    _156 = 2,
-    _111 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.zombie = {
-  is_extra = true,
-  ordering = 20100,
-  breed_power = 100,
-  male_image = 167,
-  female_image = 167,
-  min_age = 10,
-  max_age = 209,
-  height = 160,
-  male_ratio = 50,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  skills = {
-    _2 = 120,
-    _3 = 100,
-    _10 = 10,
-    _11 = 8,
-    _12 = 2,
-    _13 = 2,
-    _14 = 1,
-    _15 = 6,
-    _16 = 6,
-    _17 = 2,
-    _18 = 45,
-    _106 = 3,
-    _184 = 3,
-    _185 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 6, 7, 8},
-}
-race.elea = {
-  is_extra = false,
-  ordering = 10060,
-  breed_power = 120,
-  male_image = 3,
-  female_image = 4,
-  min_age = 10,
-  max_age = 209,
-  height = 180,
-  male_ratio = 53,
-  skills = {
-    _2 = 90,
-    _3 = 110,
-    _10 = 6,
-    _11 = 6,
-    _12 = 8,
-    _13 = 8,
-    _14 = 8,
-    _15 = 12,
-    _16 = 13,
-    _17 = 10,
-    _18 = 70,
-    _106 = 2,
-    _164 = 5,
-    _172 = 2,
-    _165 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.rabbit = {
-  is_extra = true,
-  ordering = 20110,
-  breed_power = 800,
-  male_image = 169,
-  female_image = 169,
-  min_age = 2,
-  max_age = 6,
-  height = 40,
-  male_ratio = 53,
-  skills = {
-    _2 = 90,
-    _3 = 80,
-    _10 = 5,
-    _11 = 4,
-    _12 = 6,
-    _13 = 11,
-    _14 = 5,
-    _15 = 4,
-    _16 = 4,
-    _17 = 10,
-    _18 = 100,
-    _106 = 1,
-    _301 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 7},
-}
-race.sheep = {
-  is_extra = true,
-  ordering = 20120,
-  breed_power = 1000,
-  male_image = 170,
-  female_image = 170,
-  min_age = 2,
-  max_age = 6,
-  height = 150,
-  male_ratio = 53,
-  skills = {
-    _2 = 130,
-    _3 = 80,
-    _10 = 6,
-    _11 = 6,
-    _12 = 5,
-    _13 = 6,
-    _14 = 4,
-    _15 = 5,
-    _16 = 4,
-    _17 = 8,
-    _18 = 60,
-    _106 = 1,
-    _154 = 3,
-    _161 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 7, 9},
-}
-race.frog = {
-  is_extra = true,
-  ordering = 20130,
-  breed_power = 600,
-  male_image = 172,
-  female_image = 172,
-  min_age = 2,
-  max_age = 6,
-  height = 10,
-  male_ratio = 53,
-  skills = {
-    _2 = 80,
-    _3 = 80,
-    _10 = 6,
-    _11 = 6,
-    _12 = 6,
-    _13 = 8,
-    _14 = 5,
-    _15 = 5,
-    _16 = 5,
-    _17 = 6,
-    _18 = 60,
-    _106 = 1,
-    _183 = 3,
-    _160 = 2,
-  },
-  body_parts = {4},
-}
-race.centipede = {
-  is_extra = true,
-  ordering = 20140,
-  breed_power = 400,
-  male_image = 173,
-  female_image = 173,
-  min_age = 10,
-  max_age = 19,
-  height = 10,
-  male_ratio = 53,
-  skills = {
-    _2 = 90,
-    _3 = 80,
-    _10 = 6,
-    _11 = 5,
-    _12 = 5,
-    _13 = 6,
-    _14 = 3,
-    _15 = 2,
-    _16 = 2,
-    _17 = 2,
-    _18 = 70,
-    _106 = 1,
-    _186 = 3,
-  },
-  body_parts = {3, 6, 6},
-}
-race.snail = {
-  is_extra = false,
-  ordering = 10070,
-  breed_power = 500,
-  male_image = 174,
-  female_image = 174,
-  min_age = 10,
-  max_age = 19,
-  height = 8,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 4,
-    _11 = 5,
-    _12 = 3,
-    _13 = 6,
-    _14 = 3,
-    _15 = 4,
-    _16 = 2,
-    _17 = 4,
-    _18 = 25,
-    _106 = 1,
-    _111 = 5,
-  },
-  body_parts = {3},
-}
-race.mandrake = {
-  is_extra = true,
-  ordering = 20150,
-  breed_power = 80,
-  male_image = 175,
-  female_image = 175,
-  min_age = 10,
-  max_age = 19,
-  height = 25,
-  male_ratio = 53,
-  skills = {
-    _2 = 70,
-    _3 = 80,
-    _10 = 5,
-    _11 = 6,
-    _12 = 5,
-    _13 = 10,
-    _14 = 8,
-    _15 = 8,
-    _16 = 10,
-    _17 = 4,
-    _18 = 60,
-    _106 = 2,
-    _165 = 3,
-    _150 = 2,
-    _164 = 3,
-  },
-  body_parts = {1, 3},
-}
-race.beetle = {
-  is_extra = true,
-  ordering = 20160,
-  breed_power = 750,
-  male_image = 176,
-  female_image = 176,
-  min_age = 10,
-  max_age = 19,
-  height = 10,
-  male_ratio = 53,
-  pv_correction = 140,
-  skills = {
-    _2 = 80,
-    _3 = 80,
-    _10 = 5,
-    _11 = 6,
-    _12 = 5,
-    _13 = 7,
-    _14 = 4,
-    _15 = 5,
-    _16 = 4,
-    _17 = 3,
-    _18 = 75,
-    _106 = 2,
-    _159 = 3,
-    _157 = 3,
-  },
-  body_parts = {2},
-}
-race.mushroom = {
-  is_extra = true,
-  ordering = 20170,
-  breed_power = 440,
-  male_image = 177,
-  female_image = 177,
-  melee_attack_type = 7,
-  special_attack_type = 5,
-  min_age = 10,
-  max_age = 19,
-  height = 20,
-  male_ratio = 53,
-  skills = {
-    _2 = 50,
-    _3 = 80,
-    _10 = 5,
-    _11 = 7,
-    _12 = 4,
-    _13 = 8,
-    _14 = 4,
-    _15 = 8,
-    _16 = 10,
-    _17 = 4,
-    _18 = 60,
-    _106 = 1,
-    _177 = 3,
-    _178 = 2,
-  },
-  body_parts = {1, 2},
-}
-race.bat = {
-  is_extra = true,
-  ordering = 20180,
-  breed_power = 350,
-  male_image = 200,
-  female_image = 200,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 70,
-  male_ratio = 53,
-  dv_correction = 320,
-  skills = {
-    _2 = 10,
-    _3 = 80,
-    _10 = 3,
-    _11 = 5,
-    _12 = 9,
-    _13 = 9,
-    _14 = 4,
-    _15 = 3,
-    _16 = 6,
-    _17 = 2,
-    _18 = 140,
-    _106 = 2,
-    _187 = 3,
-  },
-  body_parts = {1},
-}
-race.ent = {
-  is_extra = true,
-  ordering = 20190,
-  breed_power = 35,
-  male_image = 201,
-  female_image = 201,
-  min_age = 10,
-  max_age = 19,
-  height = 1500,
-  male_ratio = 53,
-  skills = {
-    _2 = 170,
-    _3 = 80,
-    _10 = 9,
-    _11 = 12,
-    _12 = 4,
-    _13 = 6,
-    _14 = 2,
-    _15 = 3,
-    _16 = 6,
-    _17 = 2,
-    _18 = 50,
-    _106 = 3,
-    _154 = 2,
-    _176 = 4,
-  },
-  body_parts = {5, 6, 6, 7, 9},
-}
-race.lich = {
-  is_extra = false,
-  ordering = 10080,
-  breed_power = 25,
-  male_image = 202,
-  female_image = 202,
-  min_age = 10,
-  max_age = 19,
-  height = 180,
-  male_ratio = 53,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  dv_correction = 190,
-  pv_correction = 150,
-  skills = {
-    _2 = 80,
-    _3 = 140,
-    _10 = 5,
-    _11 = 8,
-    _12 = 7,
-    _13 = 10,
-    _14 = 10,
-    _15 = 13,
-    _16 = 15,
-    _17 = 4,
-    _18 = 70,
-    _106 = 3,
-    _155 = 5,
-    _174 = 3,
-    _172 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.hound = {
-  is_extra = true,
-  ordering = 20200,
-  breed_power = 540,
-  male_image = 203,
-  female_image = 203,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 160,
-  male_ratio = 53,
-  dv_correction = 120,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 8,
-    _11 = 9,
-    _12 = 11,
-    _13 = 10,
-    _14 = 5,
-    _15 = 4,
-    _16 = 3,
-    _17 = 3,
-    _18 = 90,
-    _106 = 2,
-    _159 = 4,
-    _183 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 9},
-}
-race.ghost = {
-  is_extra = true,
-  ordering = 20210,
-  breed_power = 30,
-  male_image = 205,
-  female_image = 205,
-  melee_attack_type = 6,
-  min_age = 10,
-  max_age = 19,
-  height = 180,
-  male_ratio = 53,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  dv_correction = 160,
-  skills = {
-    _2 = 60,
-    _3 = 80,
-    _10 = 4,
-    _11 = 8,
-    _12 = 6,
-    _13 = 14,
-    _14 = 7,
-    _15 = 6,
-    _16 = 11,
-    _17 = 11,
-    _18 = 60,
-    _106 = 2,
-    _164 = 4,
-    _174 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6},
-}
-race.spirit = {
-  is_extra = true,
-  ordering = 20220,
-  breed_power = 25,
-  male_image = 206,
-  female_image = 206,
-  min_age = 10,
-  max_age = 19,
-  height = 100,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 5,
-    _11 = 9,
-    _12 = 5,
-    _13 = 7,
-    _14 = 8,
-    _15 = 10,
-    _16 = 13,
-    _17 = 4,
-    _18 = 60,
-    _106 = 2,
-    _172 = 3,
-    _188 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6},
-}
-race.eye = {
-  is_extra = true,
-  ordering = 20230,
-  breed_power = 50,
-  male_image = 207,
-  female_image = 207,
-  melee_attack_type = 4,
-  min_age = 10,
-  max_age = 19,
-  height = 40,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 3,
-    _11 = 9,
-    _12 = 4,
-    _13 = 6,
-    _14 = 4,
-    _15 = 7,
-    _16 = 10,
-    _17 = 4,
-    _18 = 25,
-    _106 = 2,
-    _159 = 3,
-    _161 = 3,
-  },
-  body_parts = {1},
-}
-race.wyvern = {
-  is_extra = true,
-  ordering = 20240,
-  breed_power = 100,
-  male_image = 235,
-  female_image = 235,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 1600,
-  male_ratio = 53,
-  skills = {
-    _2 = 190,
-    _3 = 80,
-    _10 = 11,
-    _11 = 13,
-    _12 = 8,
-    _13 = 9,
-    _14 = 5,
-    _15 = 7,
-    _16 = 8,
-    _17 = 4,
-    _18 = 70,
-    _106 = 3,
-    _150 = 3,
-    _182 = 3,
-  },
-  body_parts = {2, 4, 6, 6},
-}
-race.wasp = {
-  is_extra = true,
-  ordering = 20250,
-  breed_power = 580,
-  male_image = 210,
-  female_image = 210,
-  melee_attack_type = 5,
-  min_age = 10,
-  max_age = 19,
-  height = 80,
-  male_ratio = 53,
-  dv_correction = 220,
-  skills = {
-    _2 = 50,
-    _3 = 80,
-    _10 = 7,
-    _11 = 6,
-    _12 = 13,
-    _13 = 11,
-    _14 = 4,
-    _15 = 6,
-    _16 = 3,
-    _17 = 2,
-    _18 = 100,
-    _106 = 2,
-    _187 = 2,
-  },
-  body_parts = {1},
-}
-race.giant = {
-  is_extra = true,
-  ordering = 20260,
-  breed_power = 60,
-  male_image = 232,
-  female_image = 232,
-  min_age = 10,
-  max_age = 19,
-  height = 1800,
-  male_ratio = 53,
-  skills = {
-    _2 = 200,
-    _3 = 80,
-    _10 = 12,
-    _11 = 14,
-    _12 = 9,
-    _13 = 4,
-    _14 = 8,
-    _15 = 5,
-    _16 = 6,
-    _17 = 3,
-    _18 = 60,
-    _106 = 4,
-    _161 = 3,
-    _174 = 2,
-    _176 = 3,
-  },
-  body_parts = {4, 3, 5, 5, 7, 9},
-}
-race.imp = {
-  is_extra = true,
-  ordering = 20270,
-  breed_power = 240,
-  male_image = 212,
-  female_image = 212,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 80,
-  male_ratio = 53,
-  dv_correction = 200,
-  skills = {
-    _2 = 70,
-    _3 = 80,
-    _10 = 5,
-    _11 = 6,
-    _12 = 8,
-    _13 = 8,
-    _14 = 8,
-    _15 = 4,
-    _16 = 7,
-    _17 = 3,
-    _18 = 70,
-    _106 = 2,
-    _165 = 3,
-    _188 = 3,
-  },
-  body_parts = {2, 4, 5, 5, 6, 6, 6},
-}
-race.hand = {
-  is_extra = true,
-  ordering = 20280,
-  breed_power = 160,
-  male_image = 213,
-  female_image = 213,
-  min_age = 10,
-  max_age = 19,
-  height = 70,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 8,
-    _11 = 9,
-    _12 = 9,
-    _13 = 4,
-    _14 = 8,
-    _15 = 6,
-    _16 = 5,
-    _17 = 2,
-    _18 = 70,
-    _106 = 2,
-    _186 = 4,
-  },
-  body_parts = {5, 5, 6, 6, 7},
-}
-race.snake = {
-  is_extra = true,
-  ordering = 20290,
-  breed_power = 430,
-  male_image = 216,
-  female_image = 216,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 50,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 6,
-    _11 = 9,
-    _12 = 8,
-    _13 = 10,
-    _14 = 4,
-    _15 = 6,
-    _16 = 6,
-    _17 = 4,
-    _18 = 70,
-    _106 = 2,
-    _157 = 4,
-  },
-  body_parts = {4},
-}
-race.drake = {
-  is_extra = true,
-  ordering = 20300,
-  breed_power = 120,
-  male_image = 233,
-  female_image = 233,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 1400,
-  male_ratio = 53,
-  pv_correction = 120,
-  skills = {
-    _2 = 160,
-    _3 = 80,
-    _10 = 11,
-    _11 = 12,
-    _12 = 11,
-    _13 = 8,
-    _14 = 6,
-    _15 = 9,
-    _16 = 8,
-    _17 = 7,
-    _18 = 85,
-    _106 = 3,
-    _182 = 3,
-    _185 = 2,
-  },
-  body_parts = {2, 4, 6, 6},
-}
-race.goblin = {
-  is_extra = false,
-  ordering = 10090,
-  breed_power = 290,
-  male_image = 220,
-  female_image = 220,
-  min_age = 10,
-  max_age = 19,
-  height = 140,
-  male_ratio = 53,
-  skills = {
-    _2 = 110,
-    _3 = 90,
-    _10 = 8,
-    _11 = 8,
-    _12 = 5,
-    _13 = 5,
-    _14 = 8,
-    _15 = 7,
-    _16 = 7,
-    _17 = 4,
-    _18 = 70,
-    _106 = 2,
-    _154 = 5,
-    _185 = 2,
-    _163 = 2,
-    _186 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.bear = {
-  is_extra = true,
-  ordering = 20310,
-  breed_power = 350,
-  male_image = 222,
-  female_image = 222,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 280,
-  male_ratio = 53,
-  skills = {
-    _2 = 160,
-    _3 = 80,
-    _10 = 9,
-    _11 = 10,
-    _12 = 5,
-    _13 = 6,
-    _14 = 6,
-    _15 = 6,
-    _16 = 3,
-    _17 = 5,
-    _18 = 70,
-    _106 = 2,
-    _154 = 3,
-    _183 = 2,
-    _186 = 3,
-  },
-  body_parts = {5, 5, 6, 6, 7, 8, 9},
-}
-race.armor = {
-  is_extra = true,
-  ordering = 20320,
-  breed_power = 40,
-  male_image = 223,
-  female_image = 223,
-  min_age = 10,
-  max_age = 19,
-  height = 550,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 300,
-  skills = {
-    _2 = 40,
-    _3 = 80,
-    _10 = 8,
-    _11 = 4,
-    _12 = 7,
-    _13 = 5,
-    _14 = 6,
-    _15 = 6,
-    _16 = 7,
-    _17 = 4,
-    _18 = 50,
-    _106 = 2,
-    _158 = 3,
-    _174 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 7, 8},
-}
-race.medusa = {
-  is_extra = true,
-  ordering = 20330,
-  breed_power = 180,
-  male_image = 224,
-  female_image = 224,
-  min_age = 10,
-  max_age = 19,
-  height = 160,
-  male_ratio = 53,
-  dv_correction = 140,
-  skills = {
-    _2 = 100,
-    _3 = 110,
-    _10 = 7,
-    _11 = 6,
-    _12 = 8,
-    _13 = 7,
-    _14 = 7,
-    _15 = 9,
-    _16 = 11,
-    _17 = 5,
-    _18 = 70,
-    _106 = 3,
-    _164 = 3,
-    _188 = 3,
-  },
-  body_parts = {4, 5, 5, 6, 6, 7, 9},
-}
-race.cupid = {
-  is_extra = true,
-  ordering = 20340,
-  breed_power = 350,
-  male_image = 225,
-  female_image = 225,
-  min_age = 10,
-  max_age = 19,
-  height = 120,
-  male_ratio = 53,
-  dv_correction = 200,
-  skills = {
-    _2 = 80,
-    _3 = 130,
-    _10 = 6,
-    _11 = 7,
-    _12 = 9,
-    _13 = 8,
-    _14 = 8,
-    _15 = 7,
-    _16 = 12,
-    _17 = 8,
-    _18 = 80,
-    _106 = 2,
-    _150 = 4,
-    _188 = 3,
-  },
-  body_parts = {2, 4, 5, 5, 6, 6, 7},
-}
-race.phantom = {
-  is_extra = true,
-  ordering = 20350,
-  breed_power = 35,
-  male_image = 226,
-  female_image = 226,
-  min_age = 10,
-  max_age = 19,
-  height = 450,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  pv_correction = 200,
-  skills = {
-    _2 = 60,
-    _3 = 90,
-    _10 = 6,
-    _11 = 6,
-    _12 = 8,
-    _13 = 7,
-    _14 = 6,
-    _15 = 6,
-    _16 = 9,
-    _17 = 4,
-    _18 = 85,
-    _106 = 2,
-    _157 = 3,
-    _175 = 3,
-  },
-  body_parts = {2, 4, 5, 5, 6, 6, 7},
-}
-race.harpy = {
-  is_extra = true,
-  ordering = 20360,
-  breed_power = 420,
-  male_image = 227,
-  female_image = 227,
-  min_age = 10,
-  max_age = 19,
-  height = 140,
-  male_ratio = 53,
-  dv_correction = 150,
-  skills = {
-    _2 = 90,
-    _3 = 100,
-    _10 = 8,
-    _11 = 6,
-    _12 = 9,
-    _13 = 9,
-    _14 = 6,
-    _15 = 8,
-    _16 = 9,
-    _17 = 4,
-    _18 = 80,
-    _106 = 2,
-    _164 = 3,
-    _174 = 2,
-  },
-  body_parts = {2, 5, 5, 6, 6, 7, 9, 9},
-}
-race.dragon = {
-  is_extra = true,
-  ordering = 20370,
-  breed_power = 20,
-  male_image = 228,
-  female_image = 228,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 2400,
-  male_ratio = 53,
-  pv_correction = 120,
-  skills = {
-    _2 = 220,
-    _3 = 80,
-    _10 = 13,
-    _11 = 15,
-    _12 = 10,
-    _13 = 9,
-    _14 = 10,
-    _15 = 9,
-    _16 = 13,
-    _17 = 9,
-    _18 = 70,
-    _106 = 3,
-    _182 = 3,
-    _179 = 3,
-  },
-  body_parts = {2, 4, 6, 6},
-}
-race.dinosaur = {
-  is_extra = true,
-  ordering = 20380,
-  breed_power = 100,
-  male_image = 389,
-  female_image = 389,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 2000,
-  male_ratio = 53,
-  pv_correction = 120,
-  skills = {
-    _2 = 140,
-    _3 = 80,
-    _10 = 8,
-    _11 = 10,
-    _12 = 6,
-    _13 = 5,
-    _14 = 4,
-    _15 = 5,
-    _16 = 6,
-    _17 = 5,
-    _18 = 120,
-    _106 = 4,
-    _182 = 3,
-    _187 = 2,
-  },
-  body_parts = {2, 4, 6, 6},
-}
-race.cerberus = {
-  is_extra = true,
-  ordering = 20390,
-  breed_power = 80,
-  male_image = 229,
-  female_image = 229,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 1200,
-  male_ratio = 53,
-  skills = {
-    _2 = 160,
-    _3 = 80,
-    _10 = 11,
-    _11 = 9,
-    _12 = 10,
-    _13 = 11,
-    _14 = 6,
-    _15 = 8,
-    _16 = 7,
-    _17 = 4,
-    _18 = 85,
-    _106 = 3,
-    _159 = 3,
-    _177 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 7, 9, 9},
-}
-race.spider = {
-  is_extra = true,
-  ordering = 20400,
-  breed_power = 560,
-  male_image = 230,
-  female_image = 230,
-  melee_attack_type = 3,
-  special_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 60,
-  male_ratio = 53,
-  dv_correction = 170,
-  skills = {
-    _2 = 50,
-    _3 = 80,
-    _10 = 5,
-    _11 = 5,
-    _12 = 9,
-    _13 = 12,
-    _14 = 5,
-    _15 = 6,
-    _16 = 8,
-    _17 = 4,
-    _18 = 120,
-    _106 = 2,
-    _157 = 3,
-    _161 = 5,
-  },
-  body_parts = {6, 6},
-}
-race.golem = {
-  is_extra = false,
-  ordering = 10100,
-  breed_power = 40,
-  male_image = 231,
-  female_image = 231,
-  min_age = 10,
-  max_age = 19,
-  height = 700,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 140,
-  skills = {
-    _2 = 150,
-    _3 = 70,
-    _10 = 10,
-    _11 = 14,
-    _12 = 4,
-    _13 = 5,
-    _14 = 4,
-    _15 = 9,
-    _16 = 6,
-    _17 = 3,
-    _18 = 45,
-    _106 = 4,
-    _153 = 5,
-    _163 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.rock = {
-  is_extra = true,
-  ordering = 20410,
-  breed_power = 200,
-  male_image = 386,
-  female_image = 386,
-  min_age = 10,
-  max_age = 19,
-  height = 500,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 200,
-  skills = {
-    _2 = 40,
-    _3 = 50,
-    _10 = 8,
-    _11 = 22,
-    _12 = 3,
-    _13 = 8,
-    _14 = 2,
-    _15 = 3,
-    _16 = 4,
-    _17 = 10,
-    _18 = 70,
-    _106 = 1,
-    _153 = 3,
-    _163 = 3,
-  },
-  body_parts = {1},
-}
-race.crab = {
-  is_extra = true,
-  ordering = 20420,
-  breed_power = 420,
-  male_image = 237,
-  female_image = 237,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 50,
-  male_ratio = 53,
-  pv_correction = 150,
-  skills = {
-    _2 = 60,
-    _3 = 80,
-    _10 = 6,
-    _11 = 6,
-    _12 = 7,
-    _13 = 7,
-    _14 = 3,
-    _15 = 4,
-    _16 = 4,
-    _17 = 3,
-    _18 = 70,
-    _106 = 2,
-    _175 = 2,
-    _168 = 3,
-  },
-  body_parts = {3, 5, 5, 6, 6, 9},
-}
-race.skeleton = {
-  is_extra = true,
-  ordering = 20430,
-  breed_power = 30,
-  male_image = 241,
-  female_image = 241,
-  min_age = 10,
-  max_age = 19,
-  height = 160,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 8,
-    _11 = 7,
-    _12 = 8,
-    _13 = 6,
-    _14 = 4,
-    _15 = 6,
-    _16 = 9,
-    _17 = 4,
-    _18 = 70,
-    _106 = 2,
-    _100 = 3,
-    _168 = 2,
-    _158 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.piece = {
-  is_extra = true,
-  ordering = 20440,
-  breed_power = 25,
-  male_image = 244,
-  female_image = 244,
-  min_age = 10,
-  max_age = 19,
-  height = 750,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 150,
-  skills = {
-    _2 = 120,
-    _3 = 80,
-    _10 = 9,
-    _11 = 7,
-    _12 = 6,
-    _13 = 4,
-    _14 = 5,
-    _15 = 9,
-    _16 = 10,
-    _17 = 4,
-    _18 = 70,
-    _106 = 2,
-    _164 = 2,
-    _150 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8},
-}
-race.cat = {
-  is_extra = true,
-  ordering = 20450,
-  breed_power = 950,
-  male_image = 253,
-  female_image = 253,
-  melee_attack_type = 1,
-  special_attack_type = 4,
-  min_age = 10,
-  max_age = 19,
-  height = 60,
-  male_ratio = 53,
-  skills = {
-    _2 = 120,
-    _3 = 120,
-    _10 = 6,
-    _11 = 7,
-    _12 = 7,
-    _13 = 7,
-    _14 = 4,
-    _15 = 4,
-    _16 = 7,
-    _17 = 12,
-    _18 = 110,
-    _106 = 2,
-    _183 = 2,
-    _187 = 3,
-    _173 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 5, 9, 9},
-}
-race.dog = {
-  is_extra = true,
-  ordering = 20460,
-  breed_power = 920,
-  male_image = 254,
-  female_image = 254,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 80,
-  male_ratio = 53,
-  skills = {
-    _2 = 120,
-    _3 = 80,
-    _10 = 7,
-    _11 = 8,
-    _12 = 7,
-    _13 = 7,
-    _14 = 8,
-    _15 = 7,
-    _16 = 7,
-    _17 = 7,
-    _18 = 85,
-    _106 = 2,
-    _153 = 3,
-    _183 = 2,
-    _159 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 9, 9},
-}
-race.roran = {
-  is_extra = true,
-  ordering = 20470,
-  breed_power = 220,
-  male_image = 4,
-  female_image = 4,
-  min_age = 10,
-  max_age = 19,
-  height = 150,
-  male_ratio = 0,
-  dv_correction = 150,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 4,
-    _11 = 6,
-    _12 = 5,
-    _13 = 7,
-    _14 = 7,
-    _15 = 6,
-    _16 = 7,
-    _17 = 10,
-    _18 = 95,
-    _106 = 2,
-    _155 = 3,
-    _150 = 4,
-    _160 = 2,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.rat = {
-  is_extra = true,
-  ordering = 20480,
-  breed_power = 1100,
-  male_image = 255,
-  female_image = 255,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 30,
-  male_ratio = 53,
-  skills = {
-    _2 = 80,
-    _3 = 80,
-    _10 = 5,
-    _11 = 4,
-    _12 = 8,
-    _13 = 10,
-    _14 = 3,
-    _15 = 3,
-    _16 = 2,
-    _17 = 3,
-    _18 = 75,
-    _106 = 1,
-    _157 = 3,
-    _161 = 2,
-  },
-  body_parts = {3, 7, 8, 9},
-}
-race.shell = {
-  is_extra = true,
-  ordering = 20490,
-  breed_power = 450,
-  male_image = 256,
-  female_image = 256,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 120,
-  male_ratio = 53,
-  pv_correction = 340,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 7,
-    _11 = 7,
-    _12 = 5,
-    _13 = 4,
-    _14 = 4,
-    _15 = 5,
-    _16 = 3,
-    _17 = 4,
-    _18 = 20,
-    _106 = 1,
-    _155 = 3,
-    _162 = 3,
-  },
-  body_parts = {9},
-}
-race.catgod = {
-  is_extra = true,
-  ordering = 20500,
-  breed_power = 5,
-  male_image = 199,
-  female_image = 199,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 120,
-  male_ratio = 0,
-  dv_correction = 250,
-  skills = {
-    _2 = 120,
-    _3 = 80,
-    _10 = 12,
-    _11 = 13,
-    _12 = 21,
-    _13 = 28,
-    _14 = 8,
-    _15 = 13,
-    _16 = 12,
-    _17 = 25,
-    _18 = 500,
-    _106 = 3,
-    _173 = 3,
-    _187 = 3,
-    _186 = 2,
-  },
-  body_parts = {1, 2, 4, 5, 6, 7, 9},
-}
-race.machinegod = {
-  is_extra = true,
-  ordering = 20510,
-  breed_power = 5,
-  male_image = 349,
-  female_image = 349,
-  min_age = 10,
-  max_age = 19,
-  height = 3000,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 150,
-  skills = {
-    _2 = 200,
-    _3 = 80,
-    _10 = 15,
-    _11 = 14,
-    _12 = 11,
-    _13 = 24,
-    _14 = 12,
-    _15 = 15,
-    _16 = 8,
-    _17 = 10,
-    _18 = 70,
-    _106 = 5,
-    _110 = 30,
-  },
-  body_parts = {1, 4, 3, 5, 5, 7, 8, 9},
-}
-race.undeadgod = {
-  is_extra = true,
-  ordering = 20520,
-  breed_power = 5,
-  male_image = 350,
-  female_image = 350,
-  min_age = 10,
-  max_age = 19,
-  height = 1500,
-  male_ratio = 53,
-  skills = {
-    _2 = 150,
-    _3 = 500,
-    _10 = 10,
-    _11 = 13,
-    _12 = 14,
-    _13 = 18,
-    _14 = 16,
-    _15 = 18,
-    _16 = 28,
-    _17 = 7,
-    _18 = 110,
-    _106 = 4,
-    _188 = 3,
-    _164 = 5,
-  },
-  body_parts = {2, 4, 3, 5, 5, 6, 6},
-}
-race.machine = {
-  is_extra = true,
-  ordering = 20530,
-  breed_power = 15,
-  male_image = 270,
-  female_image = 270,
-  special_attack_type = 6,
-  min_age = 10,
-  max_age = 19,
-  height = 240,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 80,
-    _10 = 7,
-    _11 = 10,
-    _12 = 7,
-    _13 = 9,
-    _14 = 5,
-    _15 = 12,
-    _16 = 5,
-    _17 = 6,
-    _18 = 65,
-    _106 = 2,
-    _155 = 3,
-    _158 = 3,
-    _175 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.wisp = {
-  is_extra = true,
-  ordering = 20540,
-  breed_power = 25,
-  male_image = 272,
-  female_image = 272,
-  melee_attack_type = 4,
-  min_age = 10,
-  max_age = 19,
-  height = 40,
-  male_ratio = 53,
-  skills = {
-    _2 = 150,
-    _3 = 150,
-    _10 = 5,
-    _11 = 10,
-    _12 = 10,
-    _13 = 20,
-    _14 = 10,
-    _15 = 15,
-    _16 = 15,
-    _17 = 4,
-    _18 = 35,
-    _106 = 1,
-    _188 = 3,
-    _164 = 5,
-  },
-  body_parts = {1},
-}
-race.chicken = {
-  is_extra = true,
-  ordering = 20550,
-  breed_power = 1000,
-  male_image = 274,
-  female_image = 274,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 40,
-  male_ratio = 53,
-  skills = {
-    _2 = 80,
-    _3 = 80,
-    _10 = 5,
-    _11 = 8,
-    _12 = 10,
-    _13 = 11,
-    _14 = 4,
-    _15 = 4,
-    _16 = 5,
-    _17 = 7,
-    _18 = 60,
-    _106 = 1,
-    _161 = 3,
-    _155 = 3,
-  },
-  body_parts = {1},
-}
-race.stalker = {
-  is_extra = true,
-  ordering = 20560,
-  breed_power = 25,
-  male_image = 276,
-  female_image = 276,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 19,
-  height = 180,
-  male_ratio = 53,
-  resistances = {
-    _53 = 500,
-    _58 = 500,
-    _56 = 500,
-    _50 = 80,
-  },
-  dv_correction = 130,
-  skills = {
-    _2 = 120,
-    _3 = 140,
-    _10 = 9,
-    _11 = 8,
-    _12 = 5,
-    _13 = 11,
-    _14 = 7,
-    _15 = 6,
-    _16 = 9,
-    _17 = 3,
-    _18 = 100,
-    _106 = 2,
-    _186 = 3,
-    _157 = 4,
-  },
-  body_parts = {2, 5, 5, 6, 6, 7},
-}
-race.catsister = {
-  is_extra = true,
-  ordering = 20570,
-  breed_power = 5,
-  male_image = 354,
-  female_image = 354,
-  melee_attack_type = 1,
-  min_age = 10,
-  max_age = 13,
-  height = 140,
-  male_ratio = 0,
-  skills = {
-    _2 = 30,
-    _3 = 100,
-    _10 = 7,
-    _11 = 5,
-    _12 = 13,
-    _13 = 15,
-    _14 = 8,
-    _15 = 10,
-    _16 = 13,
-    _17 = 22,
-    _18 = 200,
-    _106 = 4,
-    _167 = 6,
-    _152 = 4,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.mutant = {
-  is_extra = false,
-  ordering = 10110,
-  breed_power = 50,
-  male_image = 3,
-  female_image = 4,
-  min_age = 25,
-  max_age = 74,
-  height = 180,
-  male_ratio = 53,
-  skills = {
-    _2 = 100,
-    _3 = 100,
-    _10 = 9,
-    _11 = 5,
-    _12 = 5,
-    _13 = 7,
-    _14 = 7,
-    _15 = 9,
-    _16 = 7,
-    _17 = 1,
-    _18 = 70,
-    _106 = 2,
-    _164 = 3,
-    _154 = 4,
-  },
-  body_parts = {4, 5},
-}
-race.yeek = {
-  is_extra = true,
-  ordering = 20580,
-  breed_power = 500,
-  male_image = 378,
-  female_image = 378,
-  min_age = 15,
-  max_age = 44,
-  height = 90,
-  male_ratio = 50,
-  skills = {
-    _2 = 80,
-    _3 = 90,
-    _10 = 4,
-    _11 = 7,
-    _12 = 6,
-    _13 = 7,
-    _14 = 6,
-    _15 = 4,
-    _16 = 6,
-    _17 = 6,
-    _18 = 90,
-    _106 = 1,
-    _155 = 3,
-    _156 = 4,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.yith = {
-  is_extra = true,
-  ordering = 20590,
-  breed_power = 25,
-  male_image = 429,
-  female_image = 429,
-  melee_attack_type = 6,
-  special_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 950,
-  male_ratio = 53,
-  skills = {
-    _2 = 150,
-    _3 = 200,
-    _10 = 13,
-    _11 = 14,
-    _12 = 9,
-    _13 = 15,
-    _14 = 6,
-    _15 = 5,
-    _16 = 11,
-    _17 = 11,
-    _18 = 70,
-    _106 = 3,
-    _188 = 4,
-    _155 = 3,
-    _181 = 4,
-  },
-  body_parts = {5, 5, 5, 5, 6, 6, 6, 6},
-}
-race.servant = {
-  is_extra = true,
-  ordering = 20600,
-  breed_power = 5,
-  male_image = 1,
-  female_image = 2,
-  min_age = 100,
-  max_age = 299,
-  height = 165,
-  male_ratio = 52,
-  skills = {
-    _2 = 90,
-    _3 = 150,
-    _10 = 6,
-    _11 = 6,
-    _12 = 6,
-    _13 = 6,
-    _14 = 6,
-    _15 = 6,
-    _16 = 6,
-    _17 = 6,
-    _18 = 100,
-    _106 = 3,
-    _152 = 3,
-    _172 = 4,
-    _156 = 3,
-    _111 = 3,
-    _166 = 4,
-    _110 = 4,
-    _167 = 3,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-race.horse = {
-  is_extra = true,
-  ordering = 20610,
-  breed_power = 1000,
-  male_image = 262,
-  female_image = 262,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 250,
-  male_ratio = 53,
-  skills = {
-    _2 = 150,
-    _3 = 50,
-    _10 = 9,
-    _11 = 8,
-    _12 = 4,
-    _13 = 5,
-    _14 = 4,
-    _15 = 4,
-    _16 = 3,
-    _17 = 6,
-    _18 = 125,
-    _106 = 1,
-    _154 = 4,
-  },
-  body_parts = {4, 9, 9},
-}
-race.god = {
-  is_extra = true,
-  ordering = 20620,
-  breed_power = 1,
-  male_image = 393,
-  female_image = 393,
-  min_age = 999999,
-  max_age = 999999,
-  height = 180,
-  male_ratio = 0,
-  dv_correction = 300,
-  pv_correction = 200,
-  skills = {
-    _2 = 200,
-    _3 = 100,
-    _10 = 17,
-    _11 = 13,
-    _12 = 19,
-    _13 = 17,
-    _14 = 16,
-    _15 = 25,
-    _16 = 24,
-    _17 = 21,
-    _18 = 150,
-    _106 = 6,
-    _173 = 3,
-    _187 = 3,
-    _186 = 2,
-    _110 = 8,
-    _108 = 7,
-    _166 = 7,
-    _167 = 5,
-    _152 = 7,
-  },
-  body_parts = {5, 5},
-}
-race.quickling = {
-  is_extra = true,
-  ordering = 20630,
-  breed_power = 1,
-  male_image = 281,
-  female_image = 281,
-  min_age = 10,
-  max_age = 19,
-  height = 25,
-  male_ratio = 53,
-  resistances = {_60 = 500},
-  dv_correction = 550,
-  skills = {
-    _2 = 3,
-    _3 = 40,
-    _10 = 2,
-    _11 = 4,
-    _12 = 18,
-    _13 = 16,
-    _14 = 5,
-    _15 = 8,
-    _16 = 9,
-    _17 = 7,
-    _18 = 750,
-    _106 = 1,
-    _157 = 3,
-    _173 = 7,
-    _187 = 6,
-  },
-  body_parts = {1, 4, 3, 5, 5, 6, 6, 7, 9},
-}
-race.metal = {
-  is_extra = true,
-  ordering = 20640,
-  breed_power = 1,
-  male_image = 252,
-  female_image = 252,
-  melee_attack_type = 3,
-  min_age = 10,
-  max_age = 19,
-  height = 12,
-  male_ratio = 53,
-  resistances = {_60 = 500},
-  dv_correction = 150,
-  pv_correction = 1000,
-  skills = {
-    _2 = 1,
-    _3 = 100,
-    _10 = 4,
-    _11 = 1,
-    _12 = 3,
-    _13 = 32,
-    _14 = 2,
-    _15 = 4,
-    _16 = 4,
-    _17 = 16,
-    _18 = 640,
-    _106 = 1,
-    _164 = 4,
-    _187 = 6,
-  },
-  body_parts = {1, 4, 3},
-}
-race.bike = {
-  is_extra = true,
-  ordering = 20650,
-  breed_power = 15,
-  male_image = 471,
-  female_image = 471,
-  special_attack_type = 6,
-  min_age = 10,
-  max_age = 19,
-  height = 240,
-  is_made_of_rock = true,
-  male_ratio = 53,
-  pv_correction = 150,
-  skills = {
-    _2 = 170,
-    _3 = 60,
-    _10 = 11,
-    _11 = 13,
-    _12 = 4,
-    _13 = 3,
-    _14 = 4,
-    _15 = 5,
-    _16 = 3,
-    _17 = 11,
-    _18 = 155,
-    _106 = 2,
-    _173 = 11,
-    _158 = 3,
-    _175 = 3,
-  },
-  body_parts = {1, 4, 3, 5, 5, 8, 9, 9},
-}
-
-
-
--- For debug
-race.slug = {
-  ordering = 0,
-  pv_correction = 1000,
-  dv_correction = 1000,
-  skills = {
-    _2 = 1000,
-    _3 = 1000,
-    _10 = 100,
-    _11 = 100,
-    _12 = 100,
-    _13 = 100,
-    _14 = 100,
-    _15 = 100,
-    _16 = 100,
-    _17 = 100,
-    _18 = 500,
-    _163 = 100,
-  },
-  body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
-}
-
--- vim: et sw=2 sts=2
