+++ runtime/mods/core/data.lua
@@ -13,3 +13,4 @@ require("data/fish")
 require("data/ability")
 require("data/god")
 require("data/item_material")
+require("data/race")
