+++ runtime/mods/core/data/race.lua
@@ -1,0 +1,2389 @@
+data:define_type("race")
+data:add_multi(
+   "core.race",
+   {
+      {
+         name = "kobold",
+         is_extra = true,
+         ordering = 20010,
+         breed_power = 250,
+         male_image = 171,
+         female_image = 171,
+         min_age = 15,
+         max_age = 44,
+         height = 150,
+         male_ratio = 50,
+         skills = {
+            [2] = 90,
+            [3] = 80,
+            [10] = 8,
+            [11] = 6,
+            [12] = 5,
+            [13] = 3,
+            [14] = 2,
+            [15] = 2,
+            [16] = 4,
+            [17] = 2,
+            [18] = 70,
+            [106] = 2,
+            [300] = 3,
+            [157] = 2,
+            [174] = 3,
+         },
+         body_parts = {1, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "orc",
+         is_extra = true,
+         ordering = 20020,
+         breed_power = 300,
+         male_image = 165,
+         female_image = 165,
+         min_age = 15,
+         max_age = 44,
+         height = 150,
+         male_ratio = 51,
+         skills = {
+            [2] = 130,
+            [3] = 80,
+            [10] = 5,
+            [11] = 6,
+            [12] = 3,
+            [13] = 2,
+            [14] = 2,
+            [15] = 3,
+            [16] = 4,
+            [17] = 2,
+            [18] = 70,
+            [106] = 1,
+            [154] = 3,
+            [169] = 3,
+            [168] = 3,
+         },
+         body_parts = {1, 4, 3, 5, 5, 6, 7, 8, 9},
+      },
+      {
+         name = "troll",
+         is_extra = true,
+         ordering = 20030,
+         breed_power = 250,
+         male_image = 391,
+         female_image = 391,
+         min_age = 15,
+         max_age = 44,
+         height = 400,
+         male_ratio = 51,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 90,
+            [3] = 80,
+            [10] = 10,
+            [11] = 9,
+            [12] = 1,
+            [13] = 2,
+            [14] = 2,
+            [15] = 2,
+            [16] = 3,
+            [17] = 3,
+            [18] = 60,
+            [106] = 3,
+            [154] = 40,
+         },
+         body_parts = {1, 4, 3, 5, 5, 7, 8},
+      },
+      {
+         name = "lizardman",
+         is_extra = true,
+         ordering = 20040,
+         breed_power = 300,
+         male_image = 397,
+         female_image = 397,
+         min_age = 15,
+         max_age = 44,
+         height = 240,
+         male_ratio = 51,
+         dv_multiplier = 120,
+         skills = {
+            [2] = 120,
+            [3] = 70,
+            [10] = 7,
+            [11] = 5,
+            [12] = 6,
+            [13] = 7,
+            [14] = 4,
+            [15] = 4,
+            [16] = 5,
+            [17] = 3,
+            [18] = 80,
+            [106] = 2,
+            [104] = 3,
+            [168] = 3,
+            [173] = 2,
+         },
+         body_parts = {1, 4, 3, 5, 5, 6, 7, 8, 9},
+      },
+      {
+         name = "minotaur",
+         is_extra = true,
+         ordering = 20050,
+         breed_power = 300,
+         male_image = 398,
+         female_image = 398,
+         min_age = 15,
+         max_age = 44,
+         height = 350,
+         male_ratio = 51,
+         skills = {
+            [2] = 160,
+            [3] = 60,
+            [10] = 12,
+            [11] = 8,
+            [12] = 3,
+            [13] = 4,
+            [14] = 3,
+            [15] = 4,
+            [16] = 7,
+            [17] = 4,
+            [18] = 65,
+            [106] = 3,
+            [152] = 4,
+            [186] = 3,
+         },
+         body_parts = {1, 4, 5, 5, 6, 8, 9},
+      },
+      {
+         name = "yerles",
+         is_extra = false,
+         ordering = 10010,
+         breed_power = 220,
+         male_image = 1,
+         female_image = 2,
+         min_age = 15,
+         max_age = 34,
+         height = 165,
+         male_ratio = 52,
+         skills = {
+            [2] = 110,
+            [3] = 90,
+            [10] = 7,
+            [11] = 8,
+            [12] = 8,
+            [13] = 8,
+            [14] = 12,
+            [15] = 8,
+            [16] = 8,
+            [17] = 7,
+            [18] = 70,
+            [106] = 2,
+            [110] = 5,
+            [150] = 3,
+            [156] = 2,
+            [111] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "norland",
+         is_extra = true,
+         ordering = 20060,
+         breed_power = 220,
+         male_image = 1,
+         female_image = 2,
+         min_age = 15,
+         max_age = 34,
+         height = 170,
+         male_ratio = 52,
+         skills = {
+            [2] = 100,
+            [3] = 90,
+            [10] = 8,
+            [11] = 8,
+            [12] = 6,
+            [13] = 6,
+            [14] = 7,
+            [15] = 9,
+            [16] = 7,
+            [17] = 6,
+            [18] = 70,
+            [106] = 2,
+            [172] = 3,
+            [152] = 3,
+            [167] = 3,
+            [188] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "eulderna",
+         is_extra = false,
+         ordering = 10020,
+         breed_power = 180,
+         male_image = 5,
+         female_image = 6,
+         min_age = 16,
+         max_age = 35,
+         height = 175,
+         male_ratio = 52,
+         skills = {
+            [2] = 100,
+            [3] = 100,
+            [10] = 6,
+            [11] = 7,
+            [12] = 7,
+            [13] = 10,
+            [14] = 8,
+            [15] = 10,
+            [16] = 12,
+            [17] = 8,
+            [18] = 70,
+            [106] = 2,
+            [172] = 5,
+            [150] = 3,
+            [174] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "fairy",
+         is_extra = false,
+         ordering = 10030,
+         breed_power = 180,
+         male_image = 390,
+         female_image = 390,
+         min_age = 5,
+         max_age = 104,
+         height = 50,
+         male_ratio = 52,
+         resistances = {
+            [60] = 200,
+            [53] = 200,
+            [58] = 200,
+            [56] = 200,
+            [54] = 200,
+            [57] = 200,
+            [59] = 200,
+         },
+         dv_multiplier = 250,
+         skills = {
+            [2] = 40,
+            [3] = 130,
+            [10] = 3,
+            [11] = 4,
+            [12] = 9,
+            [13] = 10,
+            [14] = 8,
+            [15] = 8,
+            [16] = 13,
+            [17] = 12,
+            [18] = 120,
+            [106] = 1,
+            [172] = 5,
+            [300] = 3,
+            [171] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "asura",
+         is_extra = true,
+         ordering = 20070,
+         breed_power = 100,
+         male_image = 405,
+         female_image = 405,
+         min_age = 15,
+         max_age = 34,
+         height = 220,
+         male_ratio = 52,
+         dv_multiplier = 200,
+         skills = {
+            [2] = 90,
+            [3] = 70,
+            [10] = 11,
+            [11] = 9,
+            [12] = 14,
+            [13] = 6,
+            [14] = 4,
+            [15] = 4,
+            [16] = 5,
+            [17] = 5,
+            [18] = 70,
+            [106] = 2,
+            [166] = 30,
+            [187] = 6,
+            [161] = 4,
+         },
+         body_parts = {5, 5, 5, 5, 2},
+      },
+      {
+         name = "slime",
+         is_extra = true,
+         ordering = 20080,
+         breed_power = 700,
+         male_image = 168,
+         female_image = 168,
+         special_attack_type = 2,
+         min_age = 1,
+         max_age = 10,
+         height = 40,
+         male_ratio = 54,
+         skills = {
+            [2] = 80,
+            [3] = 100,
+            [10] = 4,
+            [11] = 5,
+            [12] = 7,
+            [13] = 5,
+            [14] = 6,
+            [15] = 8,
+            [16] = 4,
+            [17] = 13,
+            [18] = 55,
+            [106] = 2,
+            [173] = 2,
+            [183] = 3,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "wolf",
+         is_extra = true,
+         ordering = 20090,
+         breed_power = 800,
+         male_image = 254,
+         female_image = 254,
+         min_age = 2,
+         max_age = 11,
+         height = 100,
+         male_ratio = 55,
+         dv_multiplier = 140,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 8,
+            [11] = 8,
+            [12] = 6,
+            [13] = 11,
+            [14] = 7,
+            [15] = 8,
+            [16] = 4,
+            [17] = 6,
+            [18] = 80,
+            [106] = 2,
+            [173] = 2,
+            [187] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 7, 9},
+      },
+      {
+         name = "dwarf",
+         is_extra = false,
+         ordering = 10040,
+         breed_power = 150,
+         male_image = 66,
+         female_image = 66,
+         min_age = 20,
+         max_age = 79,
+         height = 100,
+         male_ratio = 56,
+         skills = {
+            [2] = 100,
+            [3] = 100,
+            [10] = 10,
+            [11] = 11,
+            [12] = 8,
+            [13] = 6,
+            [14] = 6,
+            [15] = 7,
+            [16] = 6,
+            [17] = 6,
+            [18] = 70,
+            [106] = 2,
+            [184] = 4,
+            [179] = 3,
+            [163] = 4,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "juere",
+         is_extra = false,
+         ordering = 10050,
+         breed_power = 210,
+         male_image = 9,
+         female_image = 10,
+         min_age = 15,
+         max_age = 44,
+         height = 165,
+         male_ratio = 50,
+         skills = {
+            [2] = 100,
+            [3] = 100,
+            [10] = 7,
+            [11] = 8,
+            [12] = 12,
+            [13] = 10,
+            [14] = 10,
+            [15] = 8,
+            [16] = 8,
+            [17] = 11,
+            [18] = 70,
+            [106] = 2,
+            [183] = 4,
+            [158] = 2,
+            [156] = 2,
+            [111] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "zombie",
+         is_extra = true,
+         ordering = 20100,
+         breed_power = 100,
+         male_image = 167,
+         female_image = 167,
+         min_age = 10,
+         max_age = 209,
+         height = 160,
+         male_ratio = 50,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         skills = {
+            [2] = 120,
+            [3] = 100,
+            [10] = 10,
+            [11] = 8,
+            [12] = 2,
+            [13] = 2,
+            [14] = 1,
+            [15] = 6,
+            [16] = 6,
+            [17] = 2,
+            [18] = 45,
+            [106] = 3,
+            [184] = 3,
+            [185] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 6, 7, 8},
+      },
+      {
+         name = "elea",
+         is_extra = false,
+         ordering = 10060,
+         breed_power = 120,
+         male_image = 3,
+         female_image = 4,
+         min_age = 10,
+         max_age = 209,
+         height = 180,
+         male_ratio = 53,
+         skills = {
+            [2] = 90,
+            [3] = 110,
+            [10] = 6,
+            [11] = 6,
+            [12] = 8,
+            [13] = 8,
+            [14] = 8,
+            [15] = 12,
+            [16] = 13,
+            [17] = 10,
+            [18] = 70,
+            [106] = 2,
+            [164] = 5,
+            [172] = 2,
+            [165] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "rabbit",
+         is_extra = true,
+         ordering = 20110,
+         breed_power = 800,
+         male_image = 169,
+         female_image = 169,
+         min_age = 2,
+         max_age = 6,
+         height = 40,
+         male_ratio = 53,
+         skills = {
+            [2] = 90,
+            [3] = 80,
+            [10] = 5,
+            [11] = 4,
+            [12] = 6,
+            [13] = 11,
+            [14] = 5,
+            [15] = 4,
+            [16] = 4,
+            [17] = 10,
+            [18] = 100,
+            [106] = 1,
+            [301] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 7},
+      },
+      {
+         name = "sheep",
+         is_extra = true,
+         ordering = 20120,
+         breed_power = 1000,
+         male_image = 170,
+         female_image = 170,
+         min_age = 2,
+         max_age = 6,
+         height = 150,
+         male_ratio = 53,
+         skills = {
+            [2] = 130,
+            [3] = 80,
+            [10] = 6,
+            [11] = 6,
+            [12] = 5,
+            [13] = 6,
+            [14] = 4,
+            [15] = 5,
+            [16] = 4,
+            [17] = 8,
+            [18] = 60,
+            [106] = 1,
+            [154] = 3,
+            [161] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 7, 9},
+      },
+      {
+         name = "frog",
+         is_extra = true,
+         ordering = 20130,
+         breed_power = 600,
+         male_image = 172,
+         female_image = 172,
+         min_age = 2,
+         max_age = 6,
+         height = 10,
+         male_ratio = 53,
+         skills = {
+            [2] = 80,
+            [3] = 80,
+            [10] = 6,
+            [11] = 6,
+            [12] = 6,
+            [13] = 8,
+            [14] = 5,
+            [15] = 5,
+            [16] = 5,
+            [17] = 6,
+            [18] = 60,
+            [106] = 1,
+            [183] = 3,
+            [160] = 2,
+         },
+         body_parts = {4},
+      },
+      {
+         name = "centipede",
+         is_extra = true,
+         ordering = 20140,
+         breed_power = 400,
+         male_image = 173,
+         female_image = 173,
+         min_age = 10,
+         max_age = 19,
+         height = 10,
+         male_ratio = 53,
+         skills = {
+            [2] = 90,
+            [3] = 80,
+            [10] = 6,
+            [11] = 5,
+            [12] = 5,
+            [13] = 6,
+            [14] = 3,
+            [15] = 2,
+            [16] = 2,
+            [17] = 2,
+            [18] = 70,
+            [106] = 1,
+            [186] = 3,
+         },
+         body_parts = {3, 6, 6},
+      },
+      {
+         name = "snail",
+         is_extra = false,
+         ordering = 10070,
+         breed_power = 500,
+         male_image = 174,
+         female_image = 174,
+         min_age = 10,
+         max_age = 19,
+         height = 8,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 4,
+            [11] = 5,
+            [12] = 3,
+            [13] = 6,
+            [14] = 3,
+            [15] = 4,
+            [16] = 2,
+            [17] = 4,
+            [18] = 25,
+            [106] = 1,
+            [111] = 5,
+         },
+         body_parts = {3},
+      },
+      {
+         name = "mandrake",
+         is_extra = true,
+         ordering = 20150,
+         breed_power = 80,
+         male_image = 175,
+         female_image = 175,
+         min_age = 10,
+         max_age = 19,
+         height = 25,
+         male_ratio = 53,
+         skills = {
+            [2] = 70,
+            [3] = 80,
+            [10] = 5,
+            [11] = 6,
+            [12] = 5,
+            [13] = 10,
+            [14] = 8,
+            [15] = 8,
+            [16] = 10,
+            [17] = 4,
+            [18] = 60,
+            [106] = 2,
+            [165] = 3,
+            [150] = 2,
+            [164] = 3,
+         },
+         body_parts = {1, 3},
+      },
+      {
+         name = "beetle",
+         is_extra = true,
+         ordering = 20160,
+         breed_power = 750,
+         male_image = 176,
+         female_image = 176,
+         min_age = 10,
+         max_age = 19,
+         height = 10,
+         male_ratio = 53,
+         pv_multiplier = 140,
+         skills = {
+            [2] = 80,
+            [3] = 80,
+            [10] = 5,
+            [11] = 6,
+            [12] = 5,
+            [13] = 7,
+            [14] = 4,
+            [15] = 5,
+            [16] = 4,
+            [17] = 3,
+            [18] = 75,
+            [106] = 2,
+            [159] = 3,
+            [157] = 3,
+         },
+         body_parts = {2},
+      },
+      {
+         name = "mushroom",
+         is_extra = true,
+         ordering = 20170,
+         breed_power = 440,
+         male_image = 177,
+         female_image = 177,
+         melee_attack_type = 7,
+         special_attack_type = 5,
+         min_age = 10,
+         max_age = 19,
+         height = 20,
+         male_ratio = 53,
+         skills = {
+            [2] = 50,
+            [3] = 80,
+            [10] = 5,
+            [11] = 7,
+            [12] = 4,
+            [13] = 8,
+            [14] = 4,
+            [15] = 8,
+            [16] = 10,
+            [17] = 4,
+            [18] = 60,
+            [106] = 1,
+            [177] = 3,
+            [178] = 2,
+         },
+         body_parts = {1, 2},
+      },
+      {
+         name = "bat",
+         is_extra = true,
+         ordering = 20180,
+         breed_power = 350,
+         male_image = 200,
+         female_image = 200,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 70,
+         male_ratio = 53,
+         dv_multiplier = 320,
+         skills = {
+            [2] = 10,
+            [3] = 80,
+            [10] = 3,
+            [11] = 5,
+            [12] = 9,
+            [13] = 9,
+            [14] = 4,
+            [15] = 3,
+            [16] = 6,
+            [17] = 2,
+            [18] = 140,
+            [106] = 2,
+            [187] = 3,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "ent",
+         is_extra = true,
+         ordering = 20190,
+         breed_power = 35,
+         male_image = 201,
+         female_image = 201,
+         min_age = 10,
+         max_age = 19,
+         height = 1500,
+         male_ratio = 53,
+         skills = {
+            [2] = 170,
+            [3] = 80,
+            [10] = 9,
+            [11] = 12,
+            [12] = 4,
+            [13] = 6,
+            [14] = 2,
+            [15] = 3,
+            [16] = 6,
+            [17] = 2,
+            [18] = 50,
+            [106] = 3,
+            [154] = 2,
+            [176] = 4,
+         },
+         body_parts = {5, 6, 6, 7, 9},
+      },
+      {
+         name = "lich",
+         is_extra = false,
+         ordering = 10080,
+         breed_power = 25,
+         male_image = 202,
+         female_image = 202,
+         min_age = 10,
+         max_age = 19,
+         height = 180,
+         male_ratio = 53,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         dv_multiplier = 190,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 80,
+            [3] = 140,
+            [10] = 5,
+            [11] = 8,
+            [12] = 7,
+            [13] = 10,
+            [14] = 10,
+            [15] = 13,
+            [16] = 15,
+            [17] = 4,
+            [18] = 70,
+            [106] = 3,
+            [155] = 5,
+            [174] = 3,
+            [172] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "hound",
+         is_extra = true,
+         ordering = 20200,
+         breed_power = 540,
+         male_image = 203,
+         female_image = 203,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 160,
+         male_ratio = 53,
+         dv_multiplier = 120,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 8,
+            [11] = 9,
+            [12] = 11,
+            [13] = 10,
+            [14] = 5,
+            [15] = 4,
+            [16] = 3,
+            [17] = 3,
+            [18] = 90,
+            [106] = 2,
+            [159] = 4,
+            [183] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 9},
+      },
+      {
+         name = "ghost",
+         is_extra = true,
+         ordering = 20210,
+         breed_power = 30,
+         male_image = 205,
+         female_image = 205,
+         melee_attack_type = 6,
+         min_age = 10,
+         max_age = 19,
+         height = 180,
+         male_ratio = 53,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         dv_multiplier = 160,
+         skills = {
+            [2] = 60,
+            [3] = 80,
+            [10] = 4,
+            [11] = 8,
+            [12] = 6,
+            [13] = 14,
+            [14] = 7,
+            [15] = 6,
+            [16] = 11,
+            [17] = 11,
+            [18] = 60,
+            [106] = 2,
+            [164] = 4,
+            [174] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6},
+      },
+      {
+         name = "spirit",
+         is_extra = true,
+         ordering = 20220,
+         breed_power = 25,
+         male_image = 206,
+         female_image = 206,
+         min_age = 10,
+         max_age = 19,
+         height = 100,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 5,
+            [11] = 9,
+            [12] = 5,
+            [13] = 7,
+            [14] = 8,
+            [15] = 10,
+            [16] = 13,
+            [17] = 4,
+            [18] = 60,
+            [106] = 2,
+            [172] = 3,
+            [188] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6},
+      },
+      {
+         name = "eye",
+         is_extra = true,
+         ordering = 20230,
+         breed_power = 50,
+         male_image = 207,
+         female_image = 207,
+         melee_attack_type = 4,
+         min_age = 10,
+         max_age = 19,
+         height = 40,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 3,
+            [11] = 9,
+            [12] = 4,
+            [13] = 6,
+            [14] = 4,
+            [15] = 7,
+            [16] = 10,
+            [17] = 4,
+            [18] = 25,
+            [106] = 2,
+            [159] = 3,
+            [161] = 3,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "wyvern",
+         is_extra = true,
+         ordering = 20240,
+         breed_power = 100,
+         male_image = 235,
+         female_image = 235,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 1600,
+         male_ratio = 53,
+         skills = {
+            [2] = 190,
+            [3] = 80,
+            [10] = 11,
+            [11] = 13,
+            [12] = 8,
+            [13] = 9,
+            [14] = 5,
+            [15] = 7,
+            [16] = 8,
+            [17] = 4,
+            [18] = 70,
+            [106] = 3,
+            [150] = 3,
+            [182] = 3,
+         },
+         body_parts = {2, 4, 6, 6},
+      },
+      {
+         name = "wasp",
+         is_extra = true,
+         ordering = 20250,
+         breed_power = 580,
+         male_image = 210,
+         female_image = 210,
+         melee_attack_type = 5,
+         min_age = 10,
+         max_age = 19,
+         height = 80,
+         male_ratio = 53,
+         dv_multiplier = 220,
+         skills = {
+            [2] = 50,
+            [3] = 80,
+            [10] = 7,
+            [11] = 6,
+            [12] = 13,
+            [13] = 11,
+            [14] = 4,
+            [15] = 6,
+            [16] = 3,
+            [17] = 2,
+            [18] = 100,
+            [106] = 2,
+            [187] = 2,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "giant",
+         is_extra = true,
+         ordering = 20260,
+         breed_power = 60,
+         male_image = 232,
+         female_image = 232,
+         min_age = 10,
+         max_age = 19,
+         height = 1800,
+         male_ratio = 53,
+         skills = {
+            [2] = 200,
+            [3] = 80,
+            [10] = 12,
+            [11] = 14,
+            [12] = 9,
+            [13] = 4,
+            [14] = 8,
+            [15] = 5,
+            [16] = 6,
+            [17] = 3,
+            [18] = 60,
+            [106] = 4,
+            [161] = 3,
+            [174] = 2,
+            [176] = 3,
+         },
+         body_parts = {4, 3, 5, 5, 7, 9},
+      },
+      {
+         name = "imp",
+         is_extra = true,
+         ordering = 20270,
+         breed_power = 240,
+         male_image = 212,
+         female_image = 212,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 80,
+         male_ratio = 53,
+         dv_multiplier = 200,
+         skills = {
+            [2] = 70,
+            [3] = 80,
+            [10] = 5,
+            [11] = 6,
+            [12] = 8,
+            [13] = 8,
+            [14] = 8,
+            [15] = 4,
+            [16] = 7,
+            [17] = 3,
+            [18] = 70,
+            [106] = 2,
+            [165] = 3,
+            [188] = 3,
+         },
+         body_parts = {2, 4, 5, 5, 6, 6, 6},
+      },
+      {
+         name = "hand",
+         is_extra = true,
+         ordering = 20280,
+         breed_power = 160,
+         male_image = 213,
+         female_image = 213,
+         min_age = 10,
+         max_age = 19,
+         height = 70,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 8,
+            [11] = 9,
+            [12] = 9,
+            [13] = 4,
+            [14] = 8,
+            [15] = 6,
+            [16] = 5,
+            [17] = 2,
+            [18] = 70,
+            [106] = 2,
+            [186] = 4,
+         },
+         body_parts = {5, 5, 6, 6, 7},
+      },
+      {
+         name = "snake",
+         is_extra = true,
+         ordering = 20290,
+         breed_power = 430,
+         male_image = 216,
+         female_image = 216,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 50,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 6,
+            [11] = 9,
+            [12] = 8,
+            [13] = 10,
+            [14] = 4,
+            [15] = 6,
+            [16] = 6,
+            [17] = 4,
+            [18] = 70,
+            [106] = 2,
+            [157] = 4,
+         },
+         body_parts = {4},
+      },
+      {
+         name = "drake",
+         is_extra = true,
+         ordering = 20300,
+         breed_power = 120,
+         male_image = 233,
+         female_image = 233,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 1400,
+         male_ratio = 53,
+         pv_multiplier = 120,
+         skills = {
+            [2] = 160,
+            [3] = 80,
+            [10] = 11,
+            [11] = 12,
+            [12] = 11,
+            [13] = 8,
+            [14] = 6,
+            [15] = 9,
+            [16] = 8,
+            [17] = 7,
+            [18] = 85,
+            [106] = 3,
+            [182] = 3,
+            [185] = 2,
+         },
+         body_parts = {2, 4, 6, 6},
+      },
+      {
+         name = "goblin",
+         is_extra = false,
+         ordering = 10090,
+         breed_power = 290,
+         male_image = 220,
+         female_image = 220,
+         min_age = 10,
+         max_age = 19,
+         height = 140,
+         male_ratio = 53,
+         skills = {
+            [2] = 110,
+            [3] = 90,
+            [10] = 8,
+            [11] = 8,
+            [12] = 5,
+            [13] = 5,
+            [14] = 8,
+            [15] = 7,
+            [16] = 7,
+            [17] = 4,
+            [18] = 70,
+            [106] = 2,
+            [154] = 5,
+            [185] = 2,
+            [163] = 2,
+            [186] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "bear",
+         is_extra = true,
+         ordering = 20310,
+         breed_power = 350,
+         male_image = 222,
+         female_image = 222,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 280,
+         male_ratio = 53,
+         skills = {
+            [2] = 160,
+            [3] = 80,
+            [10] = 9,
+            [11] = 10,
+            [12] = 5,
+            [13] = 6,
+            [14] = 6,
+            [15] = 6,
+            [16] = 3,
+            [17] = 5,
+            [18] = 70,
+            [106] = 2,
+            [154] = 3,
+            [183] = 2,
+            [186] = 3,
+         },
+         body_parts = {5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "armor",
+         is_extra = true,
+         ordering = 20320,
+         breed_power = 40,
+         male_image = 223,
+         female_image = 223,
+         min_age = 10,
+         max_age = 19,
+         height = 550,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 300,
+         skills = {
+            [2] = 40,
+            [3] = 80,
+            [10] = 8,
+            [11] = 4,
+            [12] = 7,
+            [13] = 5,
+            [14] = 6,
+            [15] = 6,
+            [16] = 7,
+            [17] = 4,
+            [18] = 50,
+            [106] = 2,
+            [158] = 3,
+            [174] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 7, 8},
+      },
+      {
+         name = "medusa",
+         is_extra = true,
+         ordering = 20330,
+         breed_power = 180,
+         male_image = 224,
+         female_image = 224,
+         min_age = 10,
+         max_age = 19,
+         height = 160,
+         male_ratio = 53,
+         dv_multiplier = 140,
+         skills = {
+            [2] = 100,
+            [3] = 110,
+            [10] = 7,
+            [11] = 6,
+            [12] = 8,
+            [13] = 7,
+            [14] = 7,
+            [15] = 9,
+            [16] = 11,
+            [17] = 5,
+            [18] = 70,
+            [106] = 3,
+            [164] = 3,
+            [188] = 3,
+         },
+         body_parts = {4, 5, 5, 6, 6, 7, 9},
+      },
+      {
+         name = "cupid",
+         is_extra = true,
+         ordering = 20340,
+         breed_power = 350,
+         male_image = 225,
+         female_image = 225,
+         min_age = 10,
+         max_age = 19,
+         height = 120,
+         male_ratio = 53,
+         dv_multiplier = 200,
+         skills = {
+            [2] = 80,
+            [3] = 130,
+            [10] = 6,
+            [11] = 7,
+            [12] = 9,
+            [13] = 8,
+            [14] = 8,
+            [15] = 7,
+            [16] = 12,
+            [17] = 8,
+            [18] = 80,
+            [106] = 2,
+            [150] = 4,
+            [188] = 3,
+         },
+         body_parts = {2, 4, 5, 5, 6, 6, 7},
+      },
+      {
+         name = "phantom",
+         is_extra = true,
+         ordering = 20350,
+         breed_power = 35,
+         male_image = 226,
+         female_image = 226,
+         min_age = 10,
+         max_age = 19,
+         height = 450,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         pv_multiplier = 200,
+         skills = {
+            [2] = 60,
+            [3] = 90,
+            [10] = 6,
+            [11] = 6,
+            [12] = 8,
+            [13] = 7,
+            [14] = 6,
+            [15] = 6,
+            [16] = 9,
+            [17] = 4,
+            [18] = 85,
+            [106] = 2,
+            [157] = 3,
+            [175] = 3,
+         },
+         body_parts = {2, 4, 5, 5, 6, 6, 7},
+      },
+      {
+         name = "harpy",
+         is_extra = true,
+         ordering = 20360,
+         breed_power = 420,
+         male_image = 227,
+         female_image = 227,
+         min_age = 10,
+         max_age = 19,
+         height = 140,
+         male_ratio = 53,
+         dv_multiplier = 150,
+         skills = {
+            [2] = 90,
+            [3] = 100,
+            [10] = 8,
+            [11] = 6,
+            [12] = 9,
+            [13] = 9,
+            [14] = 6,
+            [15] = 8,
+            [16] = 9,
+            [17] = 4,
+            [18] = 80,
+            [106] = 2,
+            [164] = 3,
+            [174] = 2,
+         },
+         body_parts = {2, 5, 5, 6, 6, 7, 9, 9},
+      },
+      {
+         name = "dragon",
+         is_extra = true,
+         ordering = 20370,
+         breed_power = 20,
+         male_image = 228,
+         female_image = 228,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 2400,
+         male_ratio = 53,
+         pv_multiplier = 120,
+         skills = {
+            [2] = 220,
+            [3] = 80,
+            [10] = 13,
+            [11] = 15,
+            [12] = 10,
+            [13] = 9,
+            [14] = 10,
+            [15] = 9,
+            [16] = 13,
+            [17] = 9,
+            [18] = 70,
+            [106] = 3,
+            [182] = 3,
+            [179] = 3,
+         },
+         body_parts = {2, 4, 6, 6},
+      },
+      {
+         name = "dinosaur",
+         is_extra = true,
+         ordering = 20380,
+         breed_power = 100,
+         male_image = 389,
+         female_image = 389,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 2000,
+         male_ratio = 53,
+         pv_multiplier = 120,
+         skills = {
+            [2] = 140,
+            [3] = 80,
+            [10] = 8,
+            [11] = 10,
+            [12] = 6,
+            [13] = 5,
+            [14] = 4,
+            [15] = 5,
+            [16] = 6,
+            [17] = 5,
+            [18] = 120,
+            [106] = 4,
+            [182] = 3,
+            [187] = 2,
+         },
+         body_parts = {2, 4, 6, 6},
+      },
+      {
+         name = "cerberus",
+         is_extra = true,
+         ordering = 20390,
+         breed_power = 80,
+         male_image = 229,
+         female_image = 229,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 1200,
+         male_ratio = 53,
+         skills = {
+            [2] = 160,
+            [3] = 80,
+            [10] = 11,
+            [11] = 9,
+            [12] = 10,
+            [13] = 11,
+            [14] = 6,
+            [15] = 8,
+            [16] = 7,
+            [17] = 4,
+            [18] = 85,
+            [106] = 3,
+            [159] = 3,
+            [177] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 7, 9, 9},
+      },
+      {
+         name = "spider",
+         is_extra = true,
+         ordering = 20400,
+         breed_power = 560,
+         male_image = 230,
+         female_image = 230,
+         melee_attack_type = 3,
+         special_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 60,
+         male_ratio = 53,
+         dv_multiplier = 170,
+         skills = {
+            [2] = 50,
+            [3] = 80,
+            [10] = 5,
+            [11] = 5,
+            [12] = 9,
+            [13] = 12,
+            [14] = 5,
+            [15] = 6,
+            [16] = 8,
+            [17] = 4,
+            [18] = 120,
+            [106] = 2,
+            [157] = 3,
+            [161] = 5,
+         },
+         body_parts = {6, 6},
+      },
+      {
+         name = "golem",
+         is_extra = false,
+         ordering = 10100,
+         breed_power = 40,
+         male_image = 231,
+         female_image = 231,
+         min_age = 10,
+         max_age = 19,
+         height = 700,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 140,
+         skills = {
+            [2] = 150,
+            [3] = 70,
+            [10] = 10,
+            [11] = 14,
+            [12] = 4,
+            [13] = 5,
+            [14] = 4,
+            [15] = 9,
+            [16] = 6,
+            [17] = 3,
+            [18] = 45,
+            [106] = 4,
+            [153] = 5,
+            [163] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "rock",
+         is_extra = true,
+         ordering = 20410,
+         breed_power = 200,
+         male_image = 386,
+         female_image = 386,
+         min_age = 10,
+         max_age = 19,
+         height = 500,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 200,
+         skills = {
+            [2] = 40,
+            [3] = 50,
+            [10] = 8,
+            [11] = 22,
+            [12] = 3,
+            [13] = 8,
+            [14] = 2,
+            [15] = 3,
+            [16] = 4,
+            [17] = 10,
+            [18] = 70,
+            [106] = 1,
+            [153] = 3,
+            [163] = 3,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "crab",
+         is_extra = true,
+         ordering = 20420,
+         breed_power = 420,
+         male_image = 237,
+         female_image = 237,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 50,
+         male_ratio = 53,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 60,
+            [3] = 80,
+            [10] = 6,
+            [11] = 6,
+            [12] = 7,
+            [13] = 7,
+            [14] = 3,
+            [15] = 4,
+            [16] = 4,
+            [17] = 3,
+            [18] = 70,
+            [106] = 2,
+            [175] = 2,
+            [168] = 3,
+         },
+         body_parts = {3, 5, 5, 6, 6, 9},
+      },
+      {
+         name = "skeleton",
+         is_extra = true,
+         ordering = 20430,
+         breed_power = 30,
+         male_image = 241,
+         female_image = 241,
+         min_age = 10,
+         max_age = 19,
+         height = 160,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 8,
+            [11] = 7,
+            [12] = 8,
+            [13] = 6,
+            [14] = 4,
+            [15] = 6,
+            [16] = 9,
+            [17] = 4,
+            [18] = 70,
+            [106] = 2,
+            [100] = 3,
+            [168] = 2,
+            [158] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "piece",
+         is_extra = true,
+         ordering = 20440,
+         breed_power = 25,
+         male_image = 244,
+         female_image = 244,
+         min_age = 10,
+         max_age = 19,
+         height = 750,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 120,
+            [3] = 80,
+            [10] = 9,
+            [11] = 7,
+            [12] = 6,
+            [13] = 4,
+            [14] = 5,
+            [15] = 9,
+            [16] = 10,
+            [17] = 4,
+            [18] = 70,
+            [106] = 2,
+            [164] = 2,
+            [150] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8},
+      },
+      {
+         name = "cat",
+         is_extra = true,
+         ordering = 20450,
+         breed_power = 950,
+         male_image = 253,
+         female_image = 253,
+         melee_attack_type = 1,
+         special_attack_type = 4,
+         min_age = 10,
+         max_age = 19,
+         height = 60,
+         male_ratio = 53,
+         skills = {
+            [2] = 120,
+            [3] = 120,
+            [10] = 6,
+            [11] = 7,
+            [12] = 7,
+            [13] = 7,
+            [14] = 4,
+            [15] = 4,
+            [16] = 7,
+            [17] = 12,
+            [18] = 110,
+            [106] = 2,
+            [183] = 2,
+            [187] = 3,
+            [173] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 5, 9, 9},
+      },
+      {
+         name = "dog",
+         is_extra = true,
+         ordering = 20460,
+         breed_power = 920,
+         male_image = 254,
+         female_image = 254,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 80,
+         male_ratio = 53,
+         skills = {
+            [2] = 120,
+            [3] = 80,
+            [10] = 7,
+            [11] = 8,
+            [12] = 7,
+            [13] = 7,
+            [14] = 8,
+            [15] = 7,
+            [16] = 7,
+            [17] = 7,
+            [18] = 85,
+            [106] = 2,
+            [153] = 3,
+            [183] = 2,
+            [159] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 9, 9},
+      },
+      {
+         name = "roran",
+         is_extra = true,
+         ordering = 20470,
+         breed_power = 220,
+         male_image = 4,
+         female_image = 4,
+         min_age = 10,
+         max_age = 19,
+         height = 150,
+         male_ratio = 0,
+         dv_multiplier = 150,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 4,
+            [11] = 6,
+            [12] = 5,
+            [13] = 7,
+            [14] = 7,
+            [15] = 6,
+            [16] = 7,
+            [17] = 10,
+            [18] = 95,
+            [106] = 2,
+            [155] = 3,
+            [150] = 4,
+            [160] = 2,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "rat",
+         is_extra = true,
+         ordering = 20480,
+         breed_power = 1100,
+         male_image = 255,
+         female_image = 255,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 30,
+         male_ratio = 53,
+         skills = {
+            [2] = 80,
+            [3] = 80,
+            [10] = 5,
+            [11] = 4,
+            [12] = 8,
+            [13] = 10,
+            [14] = 3,
+            [15] = 3,
+            [16] = 2,
+            [17] = 3,
+            [18] = 75,
+            [106] = 1,
+            [157] = 3,
+            [161] = 2,
+         },
+         body_parts = {3, 7, 8, 9},
+      },
+      {
+         name = "shell",
+         is_extra = true,
+         ordering = 20490,
+         breed_power = 450,
+         male_image = 256,
+         female_image = 256,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 120,
+         male_ratio = 53,
+         pv_multiplier = 340,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 7,
+            [11] = 7,
+            [12] = 5,
+            [13] = 4,
+            [14] = 4,
+            [15] = 5,
+            [16] = 3,
+            [17] = 4,
+            [18] = 20,
+            [106] = 1,
+            [155] = 3,
+            [162] = 3,
+         },
+         body_parts = {9},
+      },
+      {
+         name = "catgod",
+         is_extra = true,
+         ordering = 20500,
+         breed_power = 5,
+         male_image = 199,
+         female_image = 199,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 120,
+         male_ratio = 0,
+         dv_multiplier = 250,
+         skills = {
+            [2] = 120,
+            [3] = 80,
+            [10] = 12,
+            [11] = 13,
+            [12] = 21,
+            [13] = 28,
+            [14] = 8,
+            [15] = 13,
+            [16] = 12,
+            [17] = 25,
+            [18] = 500,
+            [106] = 3,
+            [173] = 3,
+            [187] = 3,
+            [186] = 2,
+         },
+         body_parts = {1, 2, 4, 5, 6, 7, 9},
+      },
+      {
+         name = "machinegod",
+         is_extra = true,
+         ordering = 20510,
+         breed_power = 5,
+         male_image = 349,
+         female_image = 349,
+         min_age = 10,
+         max_age = 19,
+         height = 3000,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 200,
+            [3] = 80,
+            [10] = 15,
+            [11] = 14,
+            [12] = 11,
+            [13] = 24,
+            [14] = 12,
+            [15] = 15,
+            [16] = 8,
+            [17] = 10,
+            [18] = 70,
+            [106] = 5,
+            [110] = 30,
+         },
+         body_parts = {1, 4, 3, 5, 5, 7, 8, 9},
+      },
+      {
+         name = "undeadgod",
+         is_extra = true,
+         ordering = 20520,
+         breed_power = 5,
+         male_image = 350,
+         female_image = 350,
+         min_age = 10,
+         max_age = 19,
+         height = 1500,
+         male_ratio = 53,
+         skills = {
+            [2] = 150,
+            [3] = 500,
+            [10] = 10,
+            [11] = 13,
+            [12] = 14,
+            [13] = 18,
+            [14] = 16,
+            [15] = 18,
+            [16] = 28,
+            [17] = 7,
+            [18] = 110,
+            [106] = 4,
+            [188] = 3,
+            [164] = 5,
+         },
+         body_parts = {2, 4, 3, 5, 5, 6, 6},
+      },
+      {
+         name = "machine",
+         is_extra = true,
+         ordering = 20530,
+         breed_power = 15,
+         male_image = 270,
+         female_image = 270,
+         special_attack_type = 6,
+         min_age = 10,
+         max_age = 19,
+         height = 240,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 80,
+            [10] = 7,
+            [11] = 10,
+            [12] = 7,
+            [13] = 9,
+            [14] = 5,
+            [15] = 12,
+            [16] = 5,
+            [17] = 6,
+            [18] = 65,
+            [106] = 2,
+            [155] = 3,
+            [158] = 3,
+            [175] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "wisp",
+         is_extra = true,
+         ordering = 20540,
+         breed_power = 25,
+         male_image = 272,
+         female_image = 272,
+         melee_attack_type = 4,
+         min_age = 10,
+         max_age = 19,
+         height = 40,
+         male_ratio = 53,
+         skills = {
+            [2] = 150,
+            [3] = 150,
+            [10] = 5,
+            [11] = 10,
+            [12] = 10,
+            [13] = 20,
+            [14] = 10,
+            [15] = 15,
+            [16] = 15,
+            [17] = 4,
+            [18] = 35,
+            [106] = 1,
+            [188] = 3,
+            [164] = 5,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "chicken",
+         is_extra = true,
+         ordering = 20550,
+         breed_power = 1000,
+         male_image = 274,
+         female_image = 274,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 40,
+         male_ratio = 53,
+         skills = {
+            [2] = 80,
+            [3] = 80,
+            [10] = 5,
+            [11] = 8,
+            [12] = 10,
+            [13] = 11,
+            [14] = 4,
+            [15] = 4,
+            [16] = 5,
+            [17] = 7,
+            [18] = 60,
+            [106] = 1,
+            [161] = 3,
+            [155] = 3,
+         },
+         body_parts = {1},
+      },
+      {
+         name = "stalker",
+         is_extra = true,
+         ordering = 20560,
+         breed_power = 25,
+         male_image = 276,
+         female_image = 276,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 19,
+         height = 180,
+         male_ratio = 53,
+         resistances = {
+            [53] = 500,
+            [58] = 500,
+            [56] = 500,
+            [50] = 80,
+         },
+         dv_multiplier = 130,
+         skills = {
+            [2] = 120,
+            [3] = 140,
+            [10] = 9,
+            [11] = 8,
+            [12] = 5,
+            [13] = 11,
+            [14] = 7,
+            [15] = 6,
+            [16] = 9,
+            [17] = 3,
+            [18] = 100,
+            [106] = 2,
+            [186] = 3,
+            [157] = 4,
+         },
+         body_parts = {2, 5, 5, 6, 6, 7},
+      },
+      {
+         name = "catsister",
+         is_extra = true,
+         ordering = 20570,
+         breed_power = 5,
+         male_image = 354,
+         female_image = 354,
+         melee_attack_type = 1,
+         min_age = 10,
+         max_age = 13,
+         height = 140,
+         male_ratio = 0,
+         skills = {
+            [2] = 30,
+            [3] = 100,
+            [10] = 7,
+            [11] = 5,
+            [12] = 13,
+            [13] = 15,
+            [14] = 8,
+            [15] = 10,
+            [16] = 13,
+            [17] = 22,
+            [18] = 200,
+            [106] = 4,
+            [167] = 6,
+            [152] = 4,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "mutant",
+         is_extra = false,
+         ordering = 10110,
+         breed_power = 50,
+         male_image = 3,
+         female_image = 4,
+         min_age = 25,
+         max_age = 74,
+         height = 180,
+         male_ratio = 53,
+         skills = {
+            [2] = 100,
+            [3] = 100,
+            [10] = 9,
+            [11] = 5,
+            [12] = 5,
+            [13] = 7,
+            [14] = 7,
+            [15] = 9,
+            [16] = 7,
+            [17] = 1,
+            [18] = 70,
+            [106] = 2,
+            [164] = 3,
+            [154] = 4,
+         },
+         body_parts = {4, 5},
+      },
+      {
+         name = "yeek",
+         is_extra = true,
+         ordering = 20580,
+         breed_power = 500,
+         male_image = 378,
+         female_image = 378,
+         min_age = 15,
+         max_age = 44,
+         height = 90,
+         male_ratio = 50,
+         skills = {
+            [2] = 80,
+            [3] = 90,
+            [10] = 4,
+            [11] = 7,
+            [12] = 6,
+            [13] = 7,
+            [14] = 6,
+            [15] = 4,
+            [16] = 6,
+            [17] = 6,
+            [18] = 90,
+            [106] = 1,
+            [155] = 3,
+            [156] = 4,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "yith",
+         is_extra = true,
+         ordering = 20590,
+         breed_power = 25,
+         male_image = 429,
+         female_image = 429,
+         melee_attack_type = 6,
+         special_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 950,
+         male_ratio = 53,
+         skills = {
+            [2] = 150,
+            [3] = 200,
+            [10] = 13,
+            [11] = 14,
+            [12] = 9,
+            [13] = 15,
+            [14] = 6,
+            [15] = 5,
+            [16] = 11,
+            [17] = 11,
+            [18] = 70,
+            [106] = 3,
+            [188] = 4,
+            [155] = 3,
+            [181] = 4,
+         },
+         body_parts = {5, 5, 5, 5, 6, 6, 6, 6},
+      },
+      {
+         name = "servant",
+         is_extra = true,
+         ordering = 20600,
+         breed_power = 5,
+         male_image = 1,
+         female_image = 2,
+         min_age = 100,
+         max_age = 299,
+         height = 165,
+         male_ratio = 52,
+         skills = {
+            [2] = 90,
+            [3] = 150,
+            [10] = 6,
+            [11] = 6,
+            [12] = 6,
+            [13] = 6,
+            [14] = 6,
+            [15] = 6,
+            [16] = 6,
+            [17] = 6,
+            [18] = 100,
+            [106] = 3,
+            [152] = 3,
+            [172] = 4,
+            [156] = 3,
+            [111] = 3,
+            [166] = 4,
+            [110] = 4,
+            [167] = 3,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+      {
+         name = "horse",
+         is_extra = true,
+         ordering = 20610,
+         breed_power = 1000,
+         male_image = 262,
+         female_image = 262,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 250,
+         male_ratio = 53,
+         skills = {
+            [2] = 150,
+            [3] = 50,
+            [10] = 9,
+            [11] = 8,
+            [12] = 4,
+            [13] = 5,
+            [14] = 4,
+            [15] = 4,
+            [16] = 3,
+            [17] = 6,
+            [18] = 125,
+            [106] = 1,
+            [154] = 4,
+         },
+         body_parts = {4, 9, 9},
+      },
+      {
+         name = "god",
+         is_extra = true,
+         ordering = 20620,
+         breed_power = 1,
+         male_image = 393,
+         female_image = 393,
+         min_age = 999999,
+         max_age = 999999,
+         height = 180,
+         male_ratio = 0,
+         dv_multiplier = 300,
+         pv_multiplier = 200,
+         skills = {
+            [2] = 200,
+            [3] = 100,
+            [10] = 17,
+            [11] = 13,
+            [12] = 19,
+            [13] = 17,
+            [14] = 16,
+            [15] = 25,
+            [16] = 24,
+            [17] = 21,
+            [18] = 150,
+            [106] = 6,
+            [173] = 3,
+            [187] = 3,
+            [186] = 2,
+            [110] = 8,
+            [108] = 7,
+            [166] = 7,
+            [167] = 5,
+            [152] = 7,
+         },
+         body_parts = {5, 5},
+      },
+      {
+         name = "quickling",
+         is_extra = true,
+         ordering = 20630,
+         breed_power = 1,
+         male_image = 281,
+         female_image = 281,
+         min_age = 10,
+         max_age = 19,
+         height = 25,
+         male_ratio = 53,
+         resistances = {[60] = 500},
+         dv_multiplier = 550,
+         skills = {
+            [2] = 3,
+            [3] = 40,
+            [10] = 2,
+            [11] = 4,
+            [12] = 18,
+            [13] = 16,
+            [14] = 5,
+            [15] = 8,
+            [16] = 9,
+            [17] = 7,
+            [18] = 750,
+            [106] = 1,
+            [157] = 3,
+            [173] = 7,
+            [187] = 6,
+         },
+         body_parts = {1, 4, 3, 5, 5, 6, 6, 7, 9},
+      },
+      {
+         name = "metal",
+         is_extra = true,
+         ordering = 20640,
+         breed_power = 1,
+         male_image = 252,
+         female_image = 252,
+         melee_attack_type = 3,
+         min_age = 10,
+         max_age = 19,
+         height = 12,
+         male_ratio = 53,
+         resistances = {[60] = 500},
+         dv_multiplier = 150,
+         pv_multiplier = 1000,
+         skills = {
+            [2] = 1,
+            [3] = 100,
+            [10] = 4,
+            [11] = 1,
+            [12] = 3,
+            [13] = 32,
+            [14] = 2,
+            [15] = 4,
+            [16] = 4,
+            [17] = 16,
+            [18] = 640,
+            [106] = 1,
+            [164] = 4,
+            [187] = 6,
+         },
+         body_parts = {1, 4, 3},
+      },
+      {
+         name = "bike",
+         is_extra = true,
+         ordering = 20650,
+         breed_power = 15,
+         male_image = 471,
+         female_image = 471,
+         special_attack_type = 6,
+         min_age = 10,
+         max_age = 19,
+         height = 240,
+         is_made_of_rock = true,
+         male_ratio = 53,
+         pv_multiplier = 150,
+         skills = {
+            [2] = 170,
+            [3] = 60,
+            [10] = 11,
+            [11] = 13,
+            [12] = 4,
+            [13] = 3,
+            [14] = 4,
+            [15] = 5,
+            [16] = 3,
+            [17] = 11,
+            [18] = 155,
+            [106] = 2,
+            [173] = 11,
+            [158] = 3,
+            [175] = 3,
+         },
+         body_parts = {1, 4, 3, 5, 5, 8, 9, 9},
+      },
+      -- For debug
+      {
+         name = "slug",
+         ordering = 0,
+         pv_multiplier = 1000,
+         dv_multiplier = 1000,
+         skills = {
+            [2] = 1000,
+            [3] = 1000,
+            [10] = 100,
+            [11] = 100,
+            [12] = 100,
+            [13] = 100,
+            [14] = 100,
+            [15] = 100,
+            [16] = 100,
+            [17] = 100,
+            [18] = 500,
+            [163] = 100,
+         },
+         body_parts = {1, 2, 4, 3, 5, 5, 6, 6, 7, 8, 9},
+      },
+})
