+++ src/data/types.hpp
@@ -9,5 +9,6 @@
 #include "types/type_item_material.hpp"
 #include "types/type_map.hpp"
 #include "types/type_music.hpp"
+#include "types/type_race.hpp"
 #include "types/type_sound.hpp"
 #include "types/type_trait.hpp"
